backend/requirements.txt
aiosqlite>=0.20.0
websockets>=13.0
numpy>=1.26.0
anthropic>=0.40.0
hid>=1.0.6
python-dotenv>=1.0.0
httpx>=0.27.0
